Create xltrange_to_pythonlist.vb
First iteration of a VBA user defined function that takes a range as an imput and outputs a string of the excel range convert to a nested list format. This can be used to convert datasets into a format that python can immediately read without the need for saving .csvs and using python libraries to import data.